Refactor: extract get_daily_sessions and related logic into a utils module.
